+++ src/RecoveryServices/RecoveryServices.Backup.Helpers/Conversions/ConversionHelpers.cs
@@ -260,6 +260,15 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Helpers
                 return null;
             }
 
+            if (azureFileSharePolicy.VaultRetentionPolicy != null && azureFileSharePolicy.VaultRetentionPolicy.GetType() !=
+                typeof(ServiceClientModel.VaultRetentionPolicy))
+            {
+                Logger.Instance.WriteDebug("Unknown VaultRetentionPolicy object received: " +
+                    azureFileSharePolicy.VaultRetentionPolicy.GetType());
+                Logger.Instance.WriteWarning(Resources.UpdateToNewAzurePowershellWarning);
+                return null;
+            }
+
             if (azureFileSharePolicy.SchedulePolicy.GetType() !=
                 typeof(ServiceClientModel.SimpleSchedulePolicy))
             {
@@ -273,11 +282,18 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Helpers
             AzureFileSharePolicy fileSharePolicyModel = policyModel as AzureFileSharePolicy;
             fileSharePolicyModel.WorkloadType = WorkloadType.AzureFiles;
             fileSharePolicyModel.BackupManagementType = BackupManagementType.AzureStorage;
-
+            
             if(azureFileSharePolicy.RetentionPolicy != null)
             {
                 fileSharePolicyModel.RetentionPolicy =
                 PolicyHelpers.GetPSLongTermRetentionPolicy((ServiceClientModel.LongTermRetentionPolicy)((ServiceClientModel.AzureFileShareProtectionPolicy)serviceClientResponse.Properties).RetentionPolicy,
+                  ((ServiceClientModel.AzureFileShareProtectionPolicy)serviceClientResponse.Properties).TimeZone, backupManagementType);
+            }
+
+            if (azureFileSharePolicy.VaultRetentionPolicy != null)
+            {
+                fileSharePolicyModel.RetentionPolicy =
+                PolicyHelpers.GetPSVaultRetentionPolicy((ServiceClientModel.VaultRetentionPolicy)((ServiceClientModel.AzureFileShareProtectionPolicy)serviceClientResponse.Properties).VaultRetentionPolicy,
                   ((ServiceClientModel.AzureFileShareProtectionPolicy)serviceClientResponse.Properties).TimeZone, backupManagementType);
             }
             
@@ -286,7 +302,7 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Helpers
                  ((ServiceClientModel.AzureFileShareProtectionPolicy)serviceClientResponse.Properties).SchedulePolicy,
                  ((ServiceClientModel.AzureFileShareProtectionPolicy)serviceClientResponse.Properties).TimeZone);
             fileSharePolicyModel.ProtectedItemsCount = ((ServiceClientModel.AzureFileShareProtectionPolicy)serviceClientResponse.
-                Properties).ProtectedItemsCount;            
+                Properties).ProtectedItemsCount;
             return policyModel;
         }