+++ src/Resources/ResourceManager/Implementation/CmdletBase/TestDeploymentCmdletBase.cs
@@ -1,7 +1,10 @@
 ﻿using System;
+using System.Collections.Generic;
+using System.Linq;
 using System.Management.Automation;
 using Microsoft.Azure.Commands.Common.Strategies;
 using Microsoft.Azure.Commands.ResourceManager.Cmdlets.Formatters;
+using Microsoft.Azure.Commands.ResourceManager.Cmdlets.NewSdkExtensions;
 using Microsoft.Azure.Commands.ResourceManager.Cmdlets.Properties;
 using Microsoft.Azure.Commands.ResourceManager.Cmdlets.SdkModels;
 using Microsoft.Azure.Commands.ResourceManager.Cmdlets.SdkModels.Deployments;
@@ -21,6 +24,24 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.Implementation.Cmdlet
                 protectedTemplateUri = TemplateUri + "?" + QueryString;
             }
             return base.GetDynamicParameters();
+        }
+
+        public void WriteOutput(TemplateValidationInfo validationInfo)
+        {
+            if (validationInfo.Errors.Count == 0)
+            {
+                var builder = new ColoredStringBuilder();
+
+                var formatter = new WhatIfOperationResultFormatter(builder);
+
+                formatter.FormatDiagnostics(validationInfo.Diagnostics, new List<PSWhatIfChange>());
+
+                WriteObject(builder.ToString());
+            }
+            else
+            {
+                WriteObject(validationInfo.Errors.Select(e => e.ToPSResourceManagerError()).ToList());
+            }
         }
 
     }