+++ src/Aks/Aks/ChangeLog.md
@@ -18,6 +18,7 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Fixed the issue that HTTP request body contains empty userAssignedIdentities object when identity type is `SystemAssigned`.
 
 ## Version 6.0.4
 * Fixed secrets exposure in example documentation.