+++ src/DataFactory/DataFactory/ChangeLog.md
@@ -18,6 +18,8 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Added support for additional MySQL connection properties.
+* Added support for Azure PostgreSQL v2, updated connection strings, and corrected Linked JSON configurations.
 
 ## Version 1.18.9
 * Added pageSize support to Salesforce V2 Source.