+++ src/DesktopVirtualization/DesktopVirtualization/ChangeLog.md
@@ -18,6 +18,8 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+
+## Version 5.4.0
 * Added top level arm object for app attach packages
 
 ## Version 4.3.2