+++ src/HealthDataAIServices/HealthDataAIServices/ChangeLog.md
@@ -18,6 +18,7 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Upgraded API version to 2024-09-20
 * Onboarded `Microsoft.HeathDataAIServices/deidServices` to private link cmdlets
 
 ## Version 0.1.0